Organizational refactor
Changed the organization of the entire GUI to be centered around one "launch_GUI.py" function in the main directory, then each tab of the GUI is in its own directory. This should increase readability and allow people to mimic the structure of the calibration parameter specification tab and use the GUI with their own models.
