--- calibration/model_funcs/__init__.py
@@ -1,2 +1,0 @@
-from . import Full_Model_stress
-from . import util_funcs
