+++ calibration_progress/create_calibration_progress_widget.py
@@ -1,0 +1,1994 @@
+# -*- coding: utf-8 -*-
+
+# Form implementation generated from reading ui file 'new_base_hort.ui'
+#
+# Created by: PyQt5 UI code generator 5.9.2
+#
+# WARNING! All changes made in this file will be lost!
+import random
+import pandas as pd
+import numpy as np
+
+
+
+from PyQt5 import QtCore, QtGui, QtWidgets
+
+from PyQt5.QtWidgets import QFileDialog, QMessageBox, QTabWidget,\
+    QGridLayout, QLabel, QPushButton, QApplication, QSpinBox, QComboBox, QTableWidget, \
+    QHBoxLayout, QAbstractItemView, QFrame, QTableWidgetItem
+from PyQt5.QtGui import QFont, QColor
+
+from matplotlib.figure import Figure
+from matplotlib.backends.backend_qt5agg import FigureCanvasQTAgg as FigureCanvas
+from matplotlib.backends.backend_qt5agg import NavigationToolbar2QT as NavigationToolbar
+from matplotlib import rcParams as rc
+
+from .phase_diagram import plot_phase_diagram
+# from .model_funcs.test_optimizer_v2 import main
+# from .model_funcs import test_optimizer_v2
+# from .utility.export import exportData
+# from .utility.import_vals import importData
+# from .utility.latex_translation import textToLatex
+
+
+
+import cgitb
+cgitb.enable(format="text")
+
+
+# class CalibrationParametersWidget(QtWidgets.QWidget):
+#     def __init__(self):
+#         super(QtWidgets.QWidget,self).__init__()
+
+#         header_font_size = 10
+#         header_weight = 75
+#         normal_font_size = 10
+#         normal_weight = 50
+
+#         lineEdit_height = 30
+#         lineEdit_width = 100
+
+#         parameter_label_height = 60
+#         parameter_label_width = 200
+
+#         spacer_item_height = lineEdit_height
+#         spacer_item_small_width = 5
+#         spacer_item_width = 30
+#         #create ''master'' spacer item to use throughout
+#         regular_spacer = QtWidgets.QSpacerItem(
+#             spacer_item_width,
+#             spacer_item_height,
+#             QtWidgets.QSizePolicy.Fixed,
+#             QtWidgets.QSizePolicy.Minimum
+#             )
+
+#         small_spacer = QtWidgets.QSpacerItem(
+#             spacer_item_small_width,
+#             spacer_item_height,
+#             QtWidgets.QSizePolicy.Fixed,
+#             QtWidgets.QSizePolicy.Minimum
+#             )
+
+
+#         self.centralwidget = QtWidgets.QWidget(self)
+#         self.centralwidget.setObjectName("centralwidget")
+#         self.gridLayout = QtWidgets.QGridLayout(self.centralwidget)
+#         self.gridLayout.setObjectName("gridLayout")
+#         # DV_flags is an array of booleans (T/F) that determines whether
+#         # a specific design variable is active
+#         # True = Active (unconstrained)
+#         # False = Inactive (constrained)
+#         #Right now, we have 17 design variables possible. Need to revisit this
+#         self.DV_flags = np.array(np.ones(shape=17),dtype=bool)
+#         self.flags = {}
+#         self.flags['modulus_flag'] = False
+#         self.flags['slope_flag'] = False
+#         self.flags['smooth_hardening_flag'] = False
+#         # Elastic properties
+#         #%% E_M
+#         self.E_M = QtWidgets.QHBoxLayout()
+#         self.create_material_property(
+#             self.E_M,
+#             "E_M",
+#             "Pa",
+#             parameter_label_width,
+#             parameter_label_height,
+#             lineEdit_width,
+#             lineEdit_height,
+#             small_spacer,
+#             regular_spacer,
+#             )
+
+#         #%% E_A
+#         self.E_A = QtWidgets.QHBoxLayout()
+#         self.create_material_property(
+#             self.E_A,
+#             "E_A",
+#             "Pa",
+#             parameter_label_width,
+#             parameter_label_height,
+#             lineEdit_width,
+#             lineEdit_height,
+#             small_spacer,
+#             regular_spacer,
+#             )
+
+#         # Transformation temperatures
+#         #%%M_s
+#         self.M_s = QtWidgets.QHBoxLayout()
+#         self.create_material_property(
+#             self.M_s,
+#             "M_s",
+#             "K",
+#             parameter_label_width,
+#             parameter_label_height,
+#             lineEdit_width,
+#             lineEdit_height,
+#             small_spacer,
+#             regular_spacer,
+#             )
+
+#         #%% A_s
+#         self.A_s = QtWidgets.QHBoxLayout()
+#         self.create_material_property(
+#             self.A_s,
+#             "A_s",
+#             "K",
+#             parameter_label_width,
+#             parameter_label_height,
+#             lineEdit_width,
+#             lineEdit_height,
+#             small_spacer,
+#             regular_spacer,
+#             )
+
+#         #%% M_s - M_f
+#         self.M_s_M_f = QtWidgets.QHBoxLayout()
+#         self.create_material_property(
+#             self.M_s_M_f,
+#             "M_s-M_f",
+#             "K",
+#             parameter_label_width,
+#             parameter_label_height,
+#             lineEdit_width,
+#             lineEdit_height,
+#             small_spacer,
+#             regular_spacer,
+#             )
+
+#         #%% A_f - A_s
+#         self.A_f_A_s = QtWidgets.QHBoxLayout()
+#         self.create_material_property(
+#             self.A_f_A_s,
+#             "A_f-A_s",
+#             "K",
+#             parameter_label_width,
+#             parameter_label_height,
+#             lineEdit_width,
+#             lineEdit_height,
+#             small_spacer,
+#             regular_spacer,
+#             )
+
+#         #%% C_A
+#         self.C_A = QtWidgets.QHBoxLayout()
+#         self.create_material_property(
+#             self.C_A,
+#             "C_A",
+#             "Pa/K",
+#             parameter_label_width,
+#             parameter_label_height,
+#             lineEdit_width,
+#             lineEdit_height,
+#             small_spacer,
+#             regular_spacer,
+#             )
+
+#         #%% C_M
+#         self.C_M = QtWidgets.QHBoxLayout()
+#         self.create_material_property(
+#             self.C_M,
+#             "C_M",
+#             "MPa/K",
+#             parameter_label_width,
+#             parameter_label_height,
+#             lineEdit_width,
+#             lineEdit_height,
+#             small_spacer,
+#             regular_spacer,
+#             )
+
+#         # Transformation strain properties
+#         #%% H_min
+#         self.H_min = QtWidgets.QHBoxLayout()
+#         self.create_material_property(
+#             self.H_min,
+#             "H_{min}",
+#             "-",
+#             parameter_label_width,
+#             parameter_label_height,
+#             lineEdit_width,
+#             lineEdit_height,
+#             small_spacer,
+#             regular_spacer,
+#             )
+#         #%% H_max - H_min
+#         self.H_max_H_min = QtWidgets.QHBoxLayout()
+#         self.create_material_property(
+#             self.H_max_H_min,
+#             "H_{max} - H_{min}",
+#             "-",
+#             parameter_label_width,
+#             parameter_label_height,
+#             lineEdit_width,
+#             lineEdit_height,
+#             small_spacer,
+#             regular_spacer,
+#             )
+#         #%% sigma_0
+#         self.sigma_0 = QtWidgets.QHBoxLayout()
+#         self.create_material_property(
+#             self.sigma_0,
+#             r"\sigma_0",
+#             "Pa",
+#             parameter_label_width,
+#             parameter_label_height,
+#             lineEdit_width,
+#             lineEdit_height,
+#             small_spacer,
+#             regular_spacer,
+#             )
+#         #%% k
+#         self.k = QtWidgets.QHBoxLayout()
+#         self.create_material_property(
+#             self.k,
+#             "k",
+#             "1/Pa",
+#             parameter_label_width,
+#             parameter_label_height,
+#             lineEdit_width,
+#             lineEdit_height,
+#             small_spacer,
+#             regular_spacer,
+#             )
+
+#         # Other properties
+#         #%% n_1
+#         self.n_1 = QtWidgets.QHBoxLayout()
+#         self.create_material_property(
+#             self.n_1,
+#             "n_1",
+#             "-",
+#             parameter_label_width,
+#             parameter_label_height,
+#             lineEdit_width,
+#             lineEdit_height,
+#             small_spacer,
+#             regular_spacer,
+#             )
+
+#         #%% n_2
+#         self.n_2 = QtWidgets.QHBoxLayout()
+#         self.create_material_property(
+#             self.n_2,
+#             "n_2",
+#             "-",
+#             parameter_label_width,
+#             parameter_label_height,
+#             lineEdit_width,
+#             lineEdit_height,
+#             small_spacer,
+#             regular_spacer,
+#             )
+
+
+#         #%% n_3
+#         self.n_3 = QtWidgets.QHBoxLayout()
+#         self.create_material_property(
+#             self.n_3,
+#             "n_3",
+#             "-",
+#             parameter_label_width,
+#             parameter_label_height,
+#             lineEdit_width,
+#             lineEdit_height,
+#             small_spacer,
+#             regular_spacer,
+#             )
+
+#         #%% n_4
+#         self.n_4 = QtWidgets.QHBoxLayout()
+#         self.create_material_property(
+#             self.n_4,
+#             "n_4",
+#             "-",
+#             parameter_label_width,
+#             parameter_label_height,
+#             lineEdit_width,
+#             lineEdit_height,
+#             small_spacer,
+#             regular_spacer,
+#             )
+#         #%% alpha
+#         self.alpha = QtWidgets.QHBoxLayout()
+#         self.create_material_property(
+#             self.alpha,
+#             r"\alpha",
+#             "1/m",
+#             parameter_label_width,
+#             parameter_label_height,
+#             lineEdit_width,
+#             lineEdit_height,
+#             small_spacer,
+#             regular_spacer,
+#             )
+
+#         #%% Left labels
+#         self.left_labels = QtWidgets.QHBoxLayout()
+#         self.left_labels.setContentsMargins(0, -1, 5, -1)
+#         self.left_labels.setObjectName("left_labels")
+#         self.parameter_label = QtWidgets.QLabel(self.centralwidget)
+
+#         self.create_label(
+#                 self.parameter_label,
+#                 "Parameter",
+#                 header_font_size,
+#                 header_weight,
+#                 parameter_label_width,
+#                 parameter_label_height,
+#                 )
+
+#         self.left_labels.addWidget(self.parameter_label)
+#         self.left_labels.addItem(small_spacer)
+#         self.lower_bound_label = QtWidgets.QLabel(self.centralwidget)
+
+#         self.create_label(
+#                 self.lower_bound_label,
+#                 "Lower Bound",
+#                 header_font_size,
+#                 header_weight,
+#                 lineEdit_width,
+#                 parameter_label_height,
+#                 )
+
+#         self.left_labels.addWidget(self.lower_bound_label)
+#         self.left_labels.addItem(regular_spacer)
+
+
+#         self.upper_bound_label = QtWidgets.QLabel(self.centralwidget)
+
+#         self.create_label(
+#                 self.upper_bound_label,
+#                 "Upper Bound",
+#                 header_font_size,
+#                 header_weight,
+#                 lineEdit_width,
+#                 parameter_label_height,
+#                 )
+
+#         self.left_labels.addWidget(self.upper_bound_label)
+
+#         self.left_labels.addItem(regular_spacer)
+
+#         self.specify_label = QtWidgets.QLabel(self.centralwidget)
+
+#         self.create_label(
+#                 self.specify_label,
+#                 "Specify?",
+#                 header_font_size,
+#                 header_weight,
+#                 parameter_label_width,
+#                 parameter_label_height,
+#                 )
+
+#         self.left_labels.addWidget(self.specify_label)
+#         self.left_labels.addItem(regular_spacer)
+#         self.value_label = QtWidgets.QLabel(self.centralwidget)
+
+#         self.create_label(
+#                 self.value_label,
+#                 "Value",
+#                 header_font_size,
+#                 header_weight,
+#                 parameter_label_width,
+#                 parameter_label_height,
+#                 )
+
+#         self.left_labels.addWidget(self.value_label)
+
+#         #%% Right labels
+#         self.right_labels = QtWidgets.QHBoxLayout()
+#         self.right_labels.setContentsMargins(0, -1, 5, -1)
+#         self.right_labels.setObjectName("right_labels")
+
+#         self.parameter_label_right = QtWidgets.QLabel(self.centralwidget)
+#         self.lower_bound_label_right = QtWidgets.QLabel(self.centralwidget)
+#         self.upper_bound_label_right = QtWidgets.QLabel(self.centralwidget)
+#         self.specify_label_right = QtWidgets.QLabel(self.centralwidget)
+#         self.value_label_right = QtWidgets.QLabel(self.centralwidget)
+
+#         self.create_label(
+#                 self.parameter_label_right,
+#                 "Parameter",
+#                 header_font_size,
+#                 header_weight,
+#                 parameter_label_width,
+#                 parameter_label_height,
+#                 )
+
+#         self.right_labels.addWidget(self.parameter_label_right)
+#         self.right_labels.addItem(small_spacer)
+
+#         self.create_label(
+#                 self.lower_bound_label_right,
+#                 "Lower Bound",
+#                 header_font_size,
+#                 header_weight,
+#                 lineEdit_width,
+#                 parameter_label_height,
+#                 )
+
+#         self.right_labels.addWidget(self.lower_bound_label_right)
+#         self.right_labels.addItem(regular_spacer)
+
+#         self.create_label(
+#                 self.upper_bound_label_right,
+#                 "Upper Bound",
+#                 header_font_size,
+#                 header_weight,
+#                 lineEdit_width,
+#                 parameter_label_height,
+#                 )
+
+#         self.right_labels.addWidget(self.upper_bound_label_right)
+#         self.right_labels.addItem(regular_spacer)
+
+#         self.create_label(
+#                 self.specify_label_right,
+#                 "Specify?",
+#                 header_font_size,
+#                 header_weight,
+#                 parameter_label_width,
+#                 parameter_label_height,
+#                 )
+
+#         self.right_labels.addWidget(self.specify_label_right)
+
+#         self.right_labels.addItem(regular_spacer)
+
+#         self.create_label(
+#                 self.value_label_right,
+#                 "Value",
+#                 header_font_size,
+#                 header_weight,
+#                 parameter_label_width,
+#                 parameter_label_height,
+#                 )
+
+#         self.right_labels.addWidget(self.value_label_right)
+
+
+#         #%% Pushbuttons
+#         self.buttons = QtWidgets.QHBoxLayout()
+#         self.buttons.setObjectName("buttons")
+
+#         font_size = 14
+#         font_weight = 75
+
+#         font = self.create_font(
+#             font_size,
+#             font_weight,
+#             bold_flag = False
+#             )
+
+#         # self.buttons.addWidget(self.pushButton_2)
+#         self.pushButton = QtWidgets.QPushButton(self.centralwidget)
+
+#         size_policy = self.set_size_policy(
+#             self.pushButton
+#             )
+#         self.pushButton.setSizePolicy(
+#             size_policy
+#             )
+
+#         self.pushButton.setMinimumSize(QtCore.QSize(200, 100))
+#         self.pushButton.setMaximumSize(QtCore.QSize(200, 100))
+#         self.pushButton.setFont(font)
+#         self.pushButton.setObjectName("pushButton")
+#         self.pushButton.setText('Calibrate')
+#         self.buttons.addWidget(self.pushButton)
+
+#         #%% Design variable flags
+#         self.prop_constraints_label = QtWidgets.QHBoxLayout()
+#         self.prop_constraints_label.setContentsMargins(-1, -1, 5, -1)
+#         self.prop_constraints_label.setObjectName("prop_params_label")
+
+#         # Header
+#         self.prop_constraints_header = QtWidgets.QLabel(self.centralwidget)
+#         self.prop_constraints_header.setText('Material property constraints')
+
+#         header_font = self.create_font(
+#             header_font_size,
+#             header_weight,
+#             bold_flag = True
+#             )
+
+#         self.prop_constraints_header.setFont(header_font)
+#         self.prop_constraints_header.setAlignment(
+#             QtCore.Qt.AlignRight|QtCore.Qt.AlignTrailing|QtCore.Qt.AlignVCenter
+#             )
+#         self.prop_constraints_header.setObjectName("prop_constraints_header")
+#         self.prop_constraints_label.addStretch()
+#         self.prop_constraints_label.addWidget(self.prop_constraints_header)
+#         self.prop_constraints_label.addStretch()
+
+#         #Modulus Flag
+
+#         self.modulus_flag = QtWidgets.QHBoxLayout()
+
+#         self.create_constraint(
+#             self.modulus_flag,
+#             "modulus_flag",
+#             r"$E_A = E_M$",
+#             text_width = 200,
+#             text_height = 43,
+#             font_weight = 75,
+#             font_size = 12,
+#             latex_flag = True
+#             )
+
+#         # Slope flag
+
+#         self.slope_flag = QtWidgets.QHBoxLayout()
+#         self.create_constraint(
+#             self.slope_flag,
+#             "slope_flag",
+#             r"$C_A = C_M$",
+#             text_width = 200,
+#             text_height = 43,
+#             font_weight = 75,
+#             font_size = 12,
+#             latex_flag = True
+#             )
+
+#         self.smooth_hardening_flag = QtWidgets.QHBoxLayout()
+
+#         self.create_constraint(
+#             self.smooth_hardening_flag,
+#             "smooth_hardening_flag",
+#             r"$n_1=n_2=n_3=n_4$",
+#             text_width = 200,
+#             text_height = 43,
+#             font_weight = 75,
+#             font_size = 12,
+#             latex_flag = True
+#             )
+
+
+#         #%% Algorithmic parameters
+
+#         self.alg_params_label = QtWidgets.QHBoxLayout()
+#         self.alg_params_label.setContentsMargins(-1, -1, 5, -1)
+#         self.alg_params_label.setObjectName("alg_params_label")
+#         # self.gen_label =
+
+#         # Header
+#         self.alg_param_header = QtWidgets.QLabel(self.centralwidget)
+#         self.alg_param_header.setText('Algorithmic Parameters')
+
+#         # self.opt_param_header.setMinimumSize(QtCore.QSize(90, 0))
+#         # self.opt_param_header.setMaximumSize(QtCore.QSize(90, 16777215))
+
+#         self.alg_param_header.setFont(header_font)
+#         self.alg_param_header.setAlignment(
+#             QtCore.Qt.AlignRight|QtCore.Qt.AlignTrailing|QtCore.Qt.AlignVCenter
+#             )
+#         self.alg_param_header.setObjectName("opt_param_header")
+#         self.alg_params_label.addStretch()
+#         self.alg_params_label.addWidget(self.alg_param_header)
+#         self.alg_params_label.addStretch()
+
+#         self.non_op_params = QtWidgets.QHBoxLayout()
+#         self.non_op_params.setContentsMargins(-1, -1, 5, -1)
+#         self.non_op_params.setObjectName("non_op_params")
+
+
+#         # delta
+#         self.delta = QtWidgets.QHBoxLayout()
+
+#         self.create_parameter(
+#                 parameter_object=self.delta,
+#                 name='delta',
+#                 text=r"$\delta$",
+#                 text_width=200,
+#                 text_height=43,
+#                 text_size=12,
+#                 text_weight=75,
+#                 lineEdit_width=lineEdit_width,
+#                 lineEdit_height=lineEdit_height,
+#                 latex_flag = True,
+#                 )
+
+#         self.sigma_cal = QtWidgets.QHBoxLayout()
+
+#         self.create_parameter(
+#                 parameter_object=self.sigma_cal,
+#                 name='sigma_cal',
+#                 text=r"$\sigma_{cal}$",
+#                 text_width=200,
+#                 text_height=43,
+#                 text_size=12,
+#                 text_weight=75,
+#                 lineEdit_width=lineEdit_width,
+#                 lineEdit_height=lineEdit_height,
+#                 latex_flag = True,
+#                 )
+
+#         #MVF_tolerance
+#         self.MVF_tol = QtWidgets.QHBoxLayout()
+
+#         self.create_parameter(
+#                 parameter_object=self.MVF_tol,
+#                 name='MVF_tol',
+#                 text="MVF Tol",
+#                 text_width=200,
+#                 text_height=43,
+#                 text_size=12,
+#                 text_weight=75,
+#                 lineEdit_width=lineEdit_width,
+#                 lineEdit_height=lineEdit_height,
+#                 latex_flag = False,
+#                 )
+
+#         #%%Optimization parameters
+#         #Might need to make this a grid with labels on the left and line edits on the right
+
+#         self.op_params_label = QtWidgets.QHBoxLayout()
+#         self.op_params_label.setContentsMargins(-1, -1, 5, -1)
+#         self.op_params_label.setObjectName("op_params")
+#         # self.gen_label =
+
+#         # Header
+#         self.opt_param_header = QtWidgets.QLabel(self.centralwidget)
+#         self.opt_param_header.setText('Optimization Parameters')
+#         # self.opt_param_header.setMinimumSize(QtCore.QSize(90, 0))
+#         # self.opt_param_header.setMaximumSize(QtCore.QSize(90, 16777215))
+
+#         self.opt_param_header.setFont(header_font)
+#         self.opt_param_header.setAlignment(
+#             QtCore.Qt.AlignRight|QtCore.Qt.AlignTrailing|QtCore.Qt.AlignVCenter
+#             )
+#         self.opt_param_header.setObjectName("opt_param_header")
+#         self.op_params_label.addStretch()
+#         self.op_params_label.addWidget(self.opt_param_header)
+#         self.op_params_label.addStretch()
+
+#         # Number of generations
+#         self.num_gens = QtWidgets.QHBoxLayout()
+#         self.num_gens.setContentsMargins(-1, -1, 5, -1)
+#         self.num_gens.setObjectName("num_gens")
+
+
+#         self.gen_label = QtWidgets.QLabel(self.centralwidget)
+#         self.gen_label.setText('Number of generations')
+
+#         size_policy = self.set_size_policy(
+#             self.gen_label
+#             )
+#         self.gen_label.setSizePolicy(
+#             size_policy
+#             )
+
+#         #self.gen_label.setMinimumSize(QtCore.QSize(90, 0))
+#         #self.gen_label.setMaximumSize(QtCore.QSize(90, 16777215))
+
+#         normal_font = self.create_font(
+#             normal_font_size,
+#             normal_weight,
+#             bold_flag = False
+#             )
+
+#         self.gen_label.setFont(normal_font)
+#         self.gen_label.setAlignment(
+#             QtCore.Qt.AlignLeft|QtCore.Qt.AlignTrailing|QtCore.Qt.AlignVCenter
+#             )
+#         self.gen_label.setObjectName("gen_label")
+#         self.num_gens.addWidget(self.gen_label)
+#         self.num_gens.addStretch()
+#         self.gen_line = QtWidgets.QLineEdit(self.centralwidget)
+
+#         size_policy = self.set_size_policy(
+#             self.gen_line
+#             )
+#         self.gen_line.setSizePolicy(
+#             size_policy
+#             )
+
+
+#         self.gen_line.setMinimumSize(QtCore.QSize(lineEdit_width, lineEdit_height))
+#         self.gen_line.setMaximumSize(QtCore.QSize(lineEdit_width, lineEdit_height))
+#         self.gen_line.setObjectName("gen_line")
+#         self.num_gens.addWidget(self.gen_line)
+
+#         self.pop_size = QtWidgets.QHBoxLayout()
+#         self.pop_size.setContentsMargins(-1, -1, 5, -1)
+#         self.pop_size.setObjectName("pop_size")
+
+#         self.pop_label = QtWidgets.QLabel(self.centralwidget)
+#         self.pop_label.setText('Population size')
+
+
+#         size_policy = self.set_size_policy(
+#             self.pop_label
+#             )
+#         self.pop_label.setSizePolicy(
+#             size_policy
+#             )
+
+#         #self.pop_label.setMinimumSize(QtCore.QSize(90, 0))
+#         #self.pop_label.setMaximumSize(QtCore.QSize(90, 16777215))
+
+#         self.pop_label.setFont(normal_font)
+#         self.pop_label.setAlignment(
+#             QtCore.Qt.AlignLeft|QtCore.Qt.AlignTrailing|QtCore.Qt.AlignVCenter
+#             )
+#         self.pop_label.setObjectName("pop_label")
+#         self.pop_size.addWidget(self.pop_label)
+#         self.pop_size.addStretch()
+#         self.pop_line = QtWidgets.QLineEdit(self.centralwidget)
+
+
+#         size_policy = self.set_size_policy(
+#             self.pop_line
+#             )
+#         self.pop_line.setSizePolicy(
+#             size_policy
+#             )
+
+
+#         self.pop_line.setMinimumSize(QtCore.QSize(lineEdit_width, lineEdit_height))
+#         self.pop_line.setMaximumSize(QtCore.QSize(lineEdit_width, lineEdit_height))
+#         self.pop_line.setObjectName("pop_line")
+
+#         self.pop_size.addWidget(self.pop_line)
+
+#         self.num_iters = QtWidgets.QHBoxLayout()
+#         self.num_iters.setContentsMargins(-1, -1, 5, -1)
+#         self.num_iters.setObjectName("num_iters")
+
+#         self.iter_label = QtWidgets.QLabel(self.centralwidget)
+#         self.iter_label.setText('Gradient-based iterations')
+
+
+#         size_policy = self.set_size_policy(
+#             self.iter_label
+#             )
+#         self.iter_label.setSizePolicy(
+#             size_policy
+#             )
+
+#         #self.iter_label.setMinimumSize(QtCore.QSize(90, 0))
+#         #self.iter_label.setMaximumSize(QtCore.QSize(90, 16777215))
+
+#         self.iter_label.setFont(normal_font)
+#         self.iter_label.setAlignment(
+#             QtCore.Qt.AlignLeft|QtCore.Qt.AlignTrailing|QtCore.Qt.AlignVCenter
+#             )
+#         self.iter_label.setObjectName("iter_label")
+#         self.num_iters.addWidget(self.iter_label)
+#         self.num_iters.addStretch()
+#         self.iter_line = QtWidgets.QLineEdit(self.centralwidget)
+
+
+#         size_policy = self.set_size_policy(
+#             self.iter_line
+#             )
+#         self.iter_line.setSizePolicy(
+#             size_policy
+#             )
+
+#         self.iter_line.setMinimumSize(QtCore.QSize(lineEdit_width, lineEdit_height))
+#         self.iter_line.setMaximumSize(QtCore.QSize(lineEdit_width, lineEdit_height))
+#         self.iter_line.setObjectName("iter_line")
+
+#         self.num_iters.addWidget(self.iter_line)
+
+
+#         #%% Grid layout commands
+#         #Will need to figure out gridLayout at a later date.
+#         #Elastic properties
+#         self.gridLayout.addLayout(self.E_M, 1, 0, 1, 1)
+#         self.gridLayout.addLayout(self.E_A, 1, 1, 1, 1)
+#         #Phase diagram properties
+#         self.gridLayout.addLayout(self.M_s, 2, 0, 1, 1)
+#         self.gridLayout.addLayout(self.M_s_M_f, 2, 1, 1, 1)
+#         self.gridLayout.addLayout(self.A_s, 3, 0, 1, 1)
+#         self.gridLayout.addLayout(self.A_f_A_s, 3, 1, 1, 1)
+#         self.gridLayout.addLayout(self.C_M, 4, 0, 1, 1)
+#         self.gridLayout.addLayout(self.C_A, 4, 1, 1, 1)
+#         #Transformation strain properties
+#         self.gridLayout.addLayout(self.H_min, 5, 0, 1, 1)
+#         self.gridLayout.addLayout(self.H_max_H_min, 5, 1, 1, 1)
+#         self.gridLayout.addLayout(self.sigma_0, 6, 0, 1, 1)
+#         self.gridLayout.addLayout(self.k, 6, 1, 1, 1)
+#         #Other properties
+#         self.gridLayout.addLayout(self.n_1, 7, 0, 1, 1)
+#         self.gridLayout.addLayout(self.n_2, 7, 1, 1, 1)
+#         self.gridLayout.addLayout(self.n_3, 8, 0, 1, 1)
+#         self.gridLayout.addLayout(self.n_4, 8, 1, 1, 1)
+#         self.gridLayout.addLayout(self.alpha, 9, 0, 1, 1)
+
+#         self.gridLayout.addLayout(self.prop_constraints_label,0,2,1,1)
+#         self.gridLayout.addLayout(self.modulus_flag,1,2,1,1)
+#         self.gridLayout.addLayout(self.slope_flag,2,2,1,1)
+#         self.gridLayout.addLayout(self.smooth_hardening_flag,3,2,1,1)
+
+#         #Algorithmic parameters
+#         self.gridLayout.addLayout(self.alg_params_label,4,2,1,1)
+#         self.gridLayout.addLayout(self.delta,5,2,1,1)
+#         self.gridLayout.addLayout(self.sigma_cal,6,2,1,1)
+#         self.gridLayout.addLayout(self.MVF_tol,7,2,1,1)
+
+#         #Optimization parameters
+#         self.gridLayout.addLayout(self.op_params_label, 8, 2, 1, 1)
+#         self.gridLayout.addLayout(self.num_gens, 9, 2, 1, 1)
+#         self.gridLayout.addLayout(self.pop_size, 10, 2, 1, 1)
+#         self.gridLayout.addLayout(self.num_iters, 11, 2, 1, 1)
+#         #Buttons and Labels
+
+#         self.gridLayout.addLayout(self.buttons, 12, 2, 2, 2)
+#         self.gridLayout.addLayout(self.right_labels, 0, 1, 1, 1)
+#         self.gridLayout.addLayout(self.left_labels, 0, 0, 1, 1)
+
+
+#         #%% Alignment commands need to be after the grid layout
+
+
+
+#         self.prop_constraints_label.setAlignment(QtCore.Qt.AlignRight)
+#         self.modulus_flag.setAlignment(QtCore.Qt.AlignRight)
+#         self.slope_flag.setAlignment(QtCore.Qt.AlignRight)
+#         self.smooth_hardening_flag.setAlignment(QtCore.Qt.AlignRight)
+
+
+#         self.pop_line.setAlignment(QtCore.Qt.AlignRight)
+#         self.iter_line.setAlignment(QtCore.Qt.AlignRight)
+
+
+#         self.num_gens.setAlignment(QtCore.Qt.AlignLeft)
+#         self.pop_size.setAlignment(QtCore.Qt.AlignLeft)
+#         self.num_iters.setAlignment(QtCore.Qt.AlignLeft)
+
+#         #%% Signals and connections
+#         # self.retranslateUi(MainWindow)
+#         # INITIAL CONDITIONS
+#         self.loadDefaults()
+
+#         # QtCore.QMetaObject.connectSlotsByName(MainWindow)
+
+#         # SIGNALS
+#         # self.pushButton_3.clicked.connect(self.importVals)
+#         # self.pushButton_2.clicked.connect(self.guess)
+#         self.pushButton.clicked.connect(self.calibrate)
+#         #self.file_button.clicked.connect(self.openFiles)
+#         #self.pushButton.clicked.connect(self.export)
+#         # self.defaults_button.clicked.connect(self.loadDefaults)
+
+#         sliders = self.centralwidget.findChildren(QtWidgets.QSlider)
+#         for slider in sliders:
+#             slider.setMinimumSize(QtCore.QSize(150, 20))
+#             slider.setMaximumSize(QtCore.QSize(150, 20))
+
+
+#         # self.tabs = QTabWidget()
+#         # self.tabs.addTab(self.centralwidget,"Material Property Calibration")
+#         # self.tab2 = QtWidgets.QWidget()
+#         # self.tabs.addTab(self.tab2,'tab 2')
+
+#         # MainWindow.setCentralWidget(self.tabs)
+
+
+
+#     # %% Functions
+#     def changeTabs(self):
+#         super().tabs.setCurrentIndex(2)
+
+#     def toggleFullScreen(self):
+#         if self.isFullScreen():
+#             self.showNormal()
+#         else:
+#             self.showFullScreen()
+
+#     def retranslateUi(self, MainWindow):
+#         _translate = QtCore.QCoreApplication.translate
+#         MainWindow.setWindowTitle(_translate("MainWindow", "MainWindow"))
+
+
+
+#     # FUNCTIONS
+#     def openFiles(self):
+#         file_browser = QtWidgets.QFileDialog()
+#         file_browser.setFileMode(QFileDialog.ExistingFiles)
+#         files = file_browser.getOpenFileNames(
+#             self.centralwidget,
+#             'Browse Files',
+#             '',
+#             'All Files(*)'
+#             )[0]
+#         text = ''
+#         for file in files:
+#             filename = file[file.rindex('/') + 1:]
+#             text += filename + '\n'
+#         self.file_labels.setText(text)
+
+
+#     def importVals(self):
+#         fileloc = QFileDialog.getOpenFileName(
+#             None,
+#             "Open MTS Data File",
+#             "",
+#             "CSV Files (*.csv)"
+#             )
+#         if fileloc:
+#             file = str(fileloc[0])
+#             params = importData(file)
+#             for i in range(1, 10):
+#                 for j in range(2):
+#                     for key in params.keys():
+#                         if key == self.gridLayout.itemAtPosition(i, j).objectName():
+#                             if "Lower Bound" in params[key]:
+#                                 self.gridLayout.itemAtPosition(i, j).itemAt(2).widget().setText(str(params[key]["Lower Bound"]))
+#                             if "Upper Bound" in params[key]:
+#                                 self.gridLayout.itemAtPosition(i, j).itemAt(4).widget().setText(str(params[key]["Upper Bound"]))
+#                             if "Guess" in params[key]:
+#                                 self.gridLayout.itemAtPosition(i, j).itemAt(6).widget().setText(str(params[key]["Guess"]))
+
+#     def uncheck(self,state):
+#         if state == QtCore.Qt.Checked:
+#             if self.E_M.sender() == self.E_M.checkBox:
+#                 print('The E_M checkbox is checked')
+#                 self.DV_flags[0] = False
+
+#             if self.E_A.sender() == self.E_A.checkBox:
+#                 print('The E_A checkbox is checked')
+#                 # if self.flags['modulus_flag'] == True:
+#                 #     self.E_A.specified_value.setEnabled(False)
+
+#                 self.DV_flags[1] = False
+
+#             if self.M_s.sender() == self.M_s.checkBox:
+#                 print('The M_s checkbox is checked')
+#                 self.DV_flags[2] = False
+
+#             if self.M_s_M_f.sender() == self.M_s_M_f.checkBox:
+#                 print('The M_s - M_f checkbox is checked')
+#                 self.DV_flags[3] = False
+
+#             if self.A_s.sender() == self.A_s.checkBox:
+#                 print('The A_s checkbox is checked')
+#                 self.DV_flags[4] = False
+
+#             if self.A_f_A_s.sender() == self.A_f_A_s.checkBox:
+#                 print('The A_f - A_s checkbox is checked')
+#                 self.DV_flags[5] = False
+
+#             if self.C_M.sender() == self.C_M.checkBox:
+#                 print('The C_M checkbox is checked')
+#                 self.DV_flags[6] = False
+
+#             if self.C_A.sender() == self.C_A.checkBox:
+#                 print('The C_A checkbox is checked')
+#                 self.DV_flags[7] = False
+
+#             if self.H_min.sender() == self.H_min.checkBox:
+#                 print('The H_min checkbox is checked')
+#                 self.DV_flags[8] = False
+
+#             if self.H_max_H_min.sender() == self.H_max_H_min.checkBox:
+#                 print('The H_max - H_min checkbox is checked')
+#                 self.DV_flags[9] = False
+
+#             if self.k.sender() == self.k.checkBox:
+#                 print('The k checkbox is checked')
+#                 self.DV_flags[10] = False
+
+#             if self.n_1.sender() == self.n_1.checkBox:
+#                 print('The n_1 checkbox is checked')
+#                 self.DV_flags[11] = False
+
+#             if self.n_2.sender() == self.n_2.checkBox:
+#                 print('The n_2 checkbox is checked')
+#                 self.DV_flags[12] = False
+
+#             if self.n_3.sender() == self.n_3.checkBox:
+#                 print('The n_3 checkbox is checked')
+#                 self.DV_flags[13] = False
+
+#             if self.n_4.sender() == self.n_4.checkBox:
+#                 print('The n_4 checkbox is checked')
+#                 self.DV_flags[14] = False
+
+#             if self.sigma_0.sender() == self.sigma_0.checkBox:
+#                 print('The sigma_0 checkbox is checked')
+#                 self.DV_flags[15] = False
+
+#             if self.alpha.sender() == self.alpha.checkBox:
+#                 print('The alpha checkbox is checked')
+#                 self.DV_flags[16] = False
+
+#             if self.modulus_flag.sender() == self.modulus_flag.checkbox:
+#                 # The modulus flag does the following:
+#                     # 1. Eliminates the design variable for E_A from the opt. problem.
+#                     # 2. Sets the modulus flag = True s.t. E_A = E_M
+#                     # 3. Clears and disables the text edit for E_A specified
+#                     #   value and bounds
+#                     # I WILL NEED TO ACCOUNT FOR THIS IN THE GET SPECIFIED VALUES
+#                     # FUNCTION
+#                 print('Modulus flag checked')
+#                 self.DV_flags[1]=False
+#                 self.flags['modulus_flag'] = True
+#                 self.E_A.checkBox.setEnabled(False)
+#                 self.E_A.minimum_bound.setEnabled(False)
+#                 self.E_A.maximum_bound.setEnabled(False)
+#                 self.E_A.specified_value.setEnabled(False)
+#                 self.E_A.specified_value.setText('')
+
+#             if self.slope_flag.sender() == self.slope_flag.checkbox:
+#                 # The slope flag does the following:
+#                     # 1. Eliminates the design variable for C_A from the opt. problem.
+#                     # 2. Sets the slope flag = True s.t. C_A = C_M
+#                     # 3. Clears and disables the text edit for C_A specified
+#                     #   value and bounds
+#                     # I WILL NEED TO ACCOUNT FOR THIS IN THE GET SPECIFIED VALUES
+#                     # FUNCTION
+#                 print('Slope flag checked')
+#                 self.DV_flags[7]=False
+#                 self.flags['slope_flag'] = True
+#                 self.C_A.checkBox.setEnabled(False)
+#                 self.C_A.minimum_bound.setEnabled(False)
+#                 self.C_A.maximum_bound.setEnabled(False)
+#                 self.C_A.specified_value.setEnabled(False)
+#                 self.C_A.specified_value.setText('')
+
+#             if self.smooth_hardening_flag.sender() == self.smooth_hardening_flag.checkbox:
+#                 # The smooth hardening flag does the following:
+#                     # 1. Eliminates the design variables for n_2, n_3
+#                     #    n_4 from the opt. problem.
+#                     # 2. Sets the smooth hardening flag = True s.t.
+#                     #    n_1 = n_2 = n_3 = n_4
+#                     # 3. Clears and disables the text edit for n_2_n_4 specified
+#                     #   value and bounds
+#                     # I WILL NEED TO ACCOUNT FOR THIS IN THE GET SPECIFIED VALUES
+#                     # FUNCTION
+#                 print('Smooth hardening flag checked')
+#                 self.DV_flags[12]=False #n_2
+#                 self.DV_flags[13]=False #n_3
+#                 self.DV_flags[14]=False #n_4
+#                 self.flags['smooth_hardening_flag'] = True
+
+#                 #n_2
+#                 self.n_2.checkBox.setEnabled(False)
+#                 self.n_2.minimum_bound.setEnabled(False)
+#                 self.n_2.maximum_bound.setEnabled(False)
+#                 self.n_2.specified_value.setEnabled(False)
+#                 self.n_2.specified_value.setText('')
+
+#                 #n_3
+#                 self.n_3.checkBox.setEnabled(False)
+#                 self.n_3.minimum_bound.setEnabled(False)
+#                 self.n_3.maximum_bound.setEnabled(False)
+#                 self.n_3.specified_value.setEnabled(False)
+#                 self.n_3.specified_value.setText('')
+
+#                 #n_4
+#                 self.n_4.checkBox.setEnabled(False)
+#                 self.n_4.minimum_bound.setEnabled(False)
+#                 self.n_4.maximum_bound.setEnabled(False)
+#                 self.n_4.specified_value.setEnabled(False)
+#                 self.n_4.specified_value.setText('')
+
+#         elif state == QtCore.Qt.Unchecked:
+#             if self.E_M.sender() == self.E_M.checkBox:
+#                 print('The E_M checkbox is unchecked')
+#                 self.DV_flags[0] = True
+
+#             if self.E_A.sender() == self.E_A.checkBox:
+#                 print('The E_A checkbox is unchecked')
+#                 self.DV_flags[1] = True
+
+#             if self.M_s.sender() == self.M_s.checkBox:
+#                 print('The M_s checkbox is unchecked')
+#                 self.DV_flags[2] = True
+
+#             if self.M_s_M_f.sender() == self.M_s_M_f.checkBox:
+#                 print('The M_s - M_f checkbox is unchecked')
+#                 self.DV_flags[3] = True
+
+#             if self.A_s.sender() == self.A_s.checkBox:
+#                 print('The A_s checkbox is unchecked')
+#                 self.DV_flags[4] = True
+
+#             if self.A_f_A_s.sender() == self.A_f_A_s.checkBox:
+#                 print('The A_f - A_s checkbox is unchecked')
+#                 self.DV_flags[5] = True
+
+#             if self.C_M.sender() == self.C_M.checkBox:
+#                 print('The C_M checkbox is unchecked')
+#                 self.DV_flags[6] = True
+
+#             if self.C_A.sender() == self.C_A.checkBox:
+#                 print('The C_A checkbox is unchecked')
+#                 self.DV_flags[7] = True
+
+#             if self.H_min.sender() == self.H_min.checkBox:
+#                 print('The H_min checkbox is unchecked')
+#                 self.DV_flags[8] = True
+
+#             if self.H_max_H_min.sender() == self.H_max_H_min.checkBox:
+#                 print('The H_max - H_min checkbox is unchecked')
+#                 self.DV_flags[9] = True
+
+#             if self.k.sender() == self.k.checkBox:
+#                 print('The k checkbox is unchecked')
+#                 self.DV_flags[10] = True
+
+#             if self.n_1.sender() == self.n_1.checkBox:
+#                 print('The n_1 checkbox is unchecked')
+#                 self.DV_flags[11] = True
+
+#             if self.n_2.sender() == self.n_2.checkBox:
+#                 print('The n_2 checkbox is unchecked')
+#                 self.DV_flags[12] = True
+
+#             if self.n_3.sender() == self.n_3.checkBox:
+#                 print('The n_3 checkbox is unchecked')
+#                 self.DV_flags[13] = True
+
+#             if self.n_4.sender() == self.n_4.checkBox:
+#                 print('The n_4 checkbox is unchecked')
+#                 self.DV_flags[14] = True
+
+#             if self.sigma_0.sender() == self.sigma_0.checkBox:
+#                 print('The sigma_0 checkbox is unchecked')
+#                 self.DV_flags[15] = True
+
+#             if self.alpha.sender() == self.alpha.checkBox:
+#                 print('The alpha checkbox is unchecked')
+#                 self.DV_flags[16] = True
+
+#             if self.modulus_flag.sender() == self.modulus_flag.checkbox:
+#                 # The modulus flag does the following:
+#                     # 1. Eliminates the design variable for E_A from the opt. problem.
+#                     # 2. Sets the modulus flag = True s.t. E_A = E_M
+#                     # 3. Clears and disables the text edit for E_A specified
+#                     #   value and bounds
+#                     # I WILL NEED TO ACCOUNT FOR THIS IN THE GET SPECIFIED VALUES
+#                     # FUNCTION
+#                 print('Modulus flag unchecked')
+#                 self.DV_flags[1]=True
+#                 self.flags['modulus_flag'] = False
+#                 self.E_A.checkBox.setEnabled(True)
+#                 self.E_A.checkBox.setChecked(False)
+#                 self.E_A.minimum_bound.setEnabled(True)
+#                 self.E_A.maximum_bound.setEnabled(True)
+#                 self.E_A.specified_value.setEnabled(False)
+#                 self.E_A.specified_value.setText('')
+
+#             if self.slope_flag.sender() == self.slope_flag.checkbox:
+#                 # The slope flag does the following:
+#                     # 1. Eliminates the design variable for C_A from the opt. problem.
+#                     # 2. Sets the slope flag = True s.t. C_A = C_M
+#                     # 3. Clears and disables the text edit for C_A specified
+#                     #   value and bounds
+#                     # I WILL NEED TO ACCOUNT FOR THIS IN THE GET SPECIFIED VALUES
+#                     # FUNCTION
+#                 print('Slope flag unchecked')
+#                 self.DV_flags[7]=True
+#                 self.flags['slope_flag'] = False
+#                 self.C_A.checkBox.setEnabled(True)
+#                 self.C_A.checkBox.setChecked(False)
+#                 self.C_A.minimum_bound.setEnabled(True)
+#                 self.C_A.maximum_bound.setEnabled(True)
+#                 self.C_A.specified_value.setEnabled(False)
+#                 self.C_A.specified_value.setText('')
+
+#             if self.smooth_hardening_flag.sender() == self.smooth_hardening_flag.checkbox:
+#                 # The smooth hardening flag does the following:
+#                     # 1. Eliminates the design variables for n_2, n_3
+#                     #    n_4 from the opt. problem.
+#                     # 2. Sets the smooth hardening flag = True s.t.
+#                     #    n_1 = n_2 = n_3 = n_4
+#                     # 3. Clears and disables the text edit for n_2_n_4 specified
+#                     #   value and bounds
+#                     # I WILL NEED TO ACCOUNT FOR THIS IN THE GET SPECIFIED VALUES
+#                     # FUNCTION
+#                 print('Smooth hardening flag unchecked')
+#                 self.DV_flags[12]=True #n_2
+#                 self.DV_flags[13]=True #n_3
+#                 self.DV_flags[14]=True #n_4
+#                 self.flags['smooth_hardening_flag'] = False
+
+#                 #n_2
+#                 self.n_2.checkBox.setEnabled(True)
+#                 self.n_2.checkBox.setChecked(False)
+#                 self.n_2.minimum_bound.setEnabled(True)
+#                 self.n_2.maximum_bound.setEnabled(True)
+#                 self.n_2.specified_value.setEnabled(False)
+#                 self.n_2.specified_value.setText('')
+
+#                 #n_3
+#                 self.n_3.checkBox.setEnabled(True)
+#                 self.n_3.checkBox.setChecked(False)
+#                 self.n_3.minimum_bound.setEnabled(True)
+#                 self.n_3.maximum_bound.setEnabled(True)
+#                 self.n_3.specified_value.setEnabled(False)
+#                 self.n_3.specified_value.setText('')
+
+#                 #n_4
+#                 self.n_4.checkBox.setEnabled(True)
+#                 self.n_4.checkBox.setChecked(False)
+#                 self.n_4.minimum_bound.setEnabled(True)
+#                 self.n_4.maximum_bound.setEnabled(True)
+#                 self.n_4.specified_value.setEnabled(False)
+#                 self.n_4.specified_value.setText('')
+
+#         print(self.DV_flags)
+
+
+#     def guess(self):
+#         invalid = False
+#         for i in range(1, 10):
+#             for j in range(2):
+#                 if i == 9 and j == 1:
+#                     continue
+#                 try:
+#                     bound = str(self.gridLayout.itemAtPosition(i, j).itemAt(4).widget().text())
+#                     upper_bound = float(bound)
+#                     bound = str(self.gridLayout.itemAtPosition(i, j).itemAt(2).widget().text())
+#                     lower_bound = float(bound)
+#                     guess = round(random.uniform(lower_bound, upper_bound), 2)
+#                     self.gridLayout.itemAtPosition(i, j).itemAt(6).widget().setText(str(guess))
+#                 except:
+#                     error = QtWidgets.QMessageBox(self.centralwidget)
+#                     error.setIcon(QMessageBox.Critical)
+#                     error.setText("Please fill all bounds first")
+#                     error.setWindowTitle("Error")
+#                     error.show()
+#                     invalid = True
+#                 if invalid:
+#                     break
+#             if invalid:
+#                 break
+
+
+#     def export(self):
+#         data = {}
+#         for i in range(1, 10):
+#             for j in range(2):
+#                 if i == 9 and j == 1:
+#                     continue
+#                 key = self.gridLayout.itemAtPosition(i, j).objectName()
+#                 data[key] = {}
+#                 try:
+#                     bound = str(self.gridLayout.itemAtPosition(i, j).itemAt(4).widget().text())
+#                     upper_bound = float(bound)
+#                     bound = str(self.gridLayout.itemAtPosition(i, j).itemAt(2).widget().text())
+#                     lower_bound = float(bound)
+#                     guess = float(self.gridLayout.itemAtPosition(i, j).itemAt(6).widget().text())
+#                     data[key]["Lower Bound"] = lower_bound
+#                     data[key]["Upper Bound"] = upper_bound
+#                     data[key]["Guess"] = guess
+#                 except:
+#                     continue
+#         exportData(data, "bounds.csv")
+
+#     def calibrate(self, parent):
+#         self.getSpecifiedValues()
+#         bounds = self.getBounds()
+#         print(self.known_values)
+#         #print(bounds)
+#         QApplication.processEvents()
+#         parent.tabs.setTabEnabled(2,True)
+#         self.changeTabs()
+
+#         error = test_optimizer_v2.main(bounds,
+#                                        self,
+#                                        parent.data_input_widget.data,
+#                                        parent.calibration_plotting)
+
+
+
+#     def loadDefaults(self):
+#         #Current order:
+#             #E_M, E_A, M_s, M_s-M_f, A_s, A_f - A_s, C_M, C_A, H_min,
+#             #H_max - H_min, k, n_1, n_2, n_3, n_4, sigma_crit, alpha
+#         lower_bound_defaults = [20E9,20E9,223.0,10.0,273.0,10.0,2E6,2E6,0.0,
+#                                 0.01,1E-8,0.01,0.01,0.01,0.01,0.0,0.0]
+#         upper_bound_defaults = [110E9,110E9,273.0,50.0,323.0,50.0,15E6,15E6,0.03,
+#                                 0.05,1E-6,1.0,1.0,1.0,1.0,60E6,1E-6]
+
+#         objects = [
+#             self.E_M,
+#             self.E_A,
+#             self.M_s,
+#             self.M_s_M_f,
+#             self.A_s,
+#             self.A_f_A_s,
+#             self.C_M,
+#             self.C_A,
+#             self.H_min,
+#             self.H_max_H_min,
+#             self.k,
+#             self.n_1,
+#             self.n_2,
+#             self.n_3,
+#             self.n_4,
+#             self.sigma_0,
+#             self.alpha,
+#             ]
+
+#         variable_formats = [
+#             "{:.2e}",
+#             "{:.2e}",
+#             "{:.0f}",
+#             "{:.0f}",
+#             "{:.0f}",
+#             "{:.0f}",
+#             "{:.2e}",
+#             "{:.2e}",
+#             "{:.2f}",
+#             "{:.2f}",
+#             "{:.2e}",
+#             "{:.2f}",
+#             "{:.2f}",
+#             "{:.2f}",
+#             "{:.2f}",
+#             "{:.2e}",
+#             "{:.2e}"
+#             ]
+
+#         for property_object,lower_bound,upper_bound,variable_format in \
+#             zip(objects,lower_bound_defaults,upper_bound_defaults,variable_formats):
+
+#             property_object.minimum_bound.setText(variable_format.format(lower_bound))
+#             property_object.maximum_bound.setText(variable_format.format(upper_bound))
+
+#         #delta
+#         self.delta.lineEdit.setText('1e-6')
+
+#         #sigma cal
+#         self.sigma_cal.lineEdit.setText('100e6')
+
+#         #MVF_tol
+#         self.MVF_tol.lineEdit.setText('1e-4')
+
+#         #number of generations
+#         self.gen_line.setText('15')
+
+#         #population size
+#         self.pop_line.setText('20')
+
+#         #number of iterations
+#         self.iter_line.setText('100')
+
+
+
+#     def updateVal(self):
+#         for i in range(1, 10):
+#             for j in range(2):
+#                 if i == 9 and j == 1:
+#                     break
+#                 upper_bound = float(self.gridLayout.itemAtPosition(i, j).itemAt(4).widget().text())
+#                 lower_bound = float(self.gridLayout.itemAtPosition(i, j).itemAt(2).widget().text())
+#                 slider_val = self.gridLayout.itemAtPosition(i, j).itemAt(3).widget().value()
+#                 guess = self.gridLayout.itemAtPosition(i, j).itemAt(6).widget()
+#                 step = (upper_bound - lower_bound)/100
+#                 guess.setText(str(slider_val * step + lower_bound))
+
+
+#     def updateSlider(self):
+#         for i in range(1, 10):
+#             for j in range(2):
+#                 if i == 9 and j == 1:
+#                     break
+#                 val = self.gridLayout.itemAtPosition(i, j).itemAt(6).widget().text()
+#                 if not val == "":
+#                     upper_bound = float(self.gridLayout.itemAtPosition(i, j).itemAt(4).widget().text())
+#                     lower_bound = float(self.gridLayout.itemAtPosition(i, j).itemAt(2).widget().text())
+#                     if float(val) < lower_bound:
+#                         print('Setting lower bound to specified value for visualization')
+#                         lower_bound = self.gridLayout.itemAtPosition(i, j).itemAt(2).widget().setText(str(val))
+#                     elif float(val) > upper_bound:
+#                         print('Setting upper bound to specified value for visualization')
+#                         upper_bound = self.gridLayout.itemAtPosition(i, j).itemAt(4).widget().setText(str(val))
+#                     upper_bound = float(self.gridLayout.itemAtPosition(i, j).itemAt(4).widget().text())
+#                     lower_bound = float(self.gridLayout.itemAtPosition(i, j).itemAt(2).widget().text())
+#                     step = (upper_bound - lower_bound)/100
+
+#                     val = float(val)
+#                     normalized_value = int(val/step-int(lower_bound)/step)
+#                     self.gridLayout.itemAtPosition(i, j).itemAt(3).widget().setValue(normalized_value)
+
+
+#     def getBounds(self):
+#         # initialize bounds list of lists
+#         bounds = []
+
+#         #E_M
+#         bounds.append([float(self.E_M.minimum_bound.text()),
+#                        float(self.E_M.maximum_bound.text())])
+
+#         #E_A
+#         # if self.flags['modulus_flag'] == False:
+#         bounds.append([float(self.E_A.minimum_bound.text()),
+#                        float(self.E_A.maximum_bound.text())])
+
+#         #M_s
+#         bounds.append([float(self.M_s.minimum_bound.text()),
+#                        float(self.M_s.maximum_bound.text())])
+
+#         #M_s-M_f
+#         bounds.append([float(self.M_s_M_f.minimum_bound.text()),
+#                        float(self.M_s_M_f.maximum_bound.text())])
+
+#         #A_s
+#         bounds.append([float(self.A_s.minimum_bound.text()),
+#                        float(self.A_s.maximum_bound.text())])
+
+#         #A_f - A_s
+#         bounds.append([float(self.A_f_A_s.minimum_bound.text()),
+#                        float(self.A_f_A_s.maximum_bound.text())])
+
+#         #C_M
+#         bounds.append([float(self.C_M.minimum_bound.text()),
+#                        float(self.C_M.maximum_bound.text())])
+
+#         #C_A
+#         # if self.flags['slope_flag'] == False:
+#         bounds.append([float(self.C_A.minimum_bound.text()),
+#                         float(self.C_A.maximum_bound.text())])
+
+#         #H_min
+#         bounds.append([float(self.H_min.minimum_bound.text()),
+#                        float(self.H_min.maximum_bound.text())])
+#         #H_max - H_min
+#         bounds.append([float(self.H_max_H_min.minimum_bound.text()),
+#                        float(self.H_max_H_min.maximum_bound.text())])
+#         #k
+#         bounds.append([float(self.k.minimum_bound.text()),
+#                        float(self.k.maximum_bound.text())])
+
+#         #n_1
+#         bounds.append([float(self.n_1.minimum_bound.text()),
+#                        float(self.n_1.maximum_bound.text())])
+
+#         #n_2
+#         # if self.flags['smooth_hardening_flag'] == False:
+#         bounds.append([float(self.n_2.minimum_bound.text()),
+#                        float(self.n_2.maximum_bound.text())])
+
+#         #n_3
+#         bounds.append([float(self.n_3.minimum_bound.text()),
+#                        float(self.n_3.maximum_bound.text())])
+
+#         #n_4
+#         bounds.append([float(self.n_4.minimum_bound.text()),
+#                        float(self.n_4.maximum_bound.text())])
+
+#         #sigma_crit
+#         bounds.append([float(self.sigma_0.minimum_bound.text()),
+#                        float(self.sigma_0.maximum_bound.text())])
+
+#         #alpha
+#         bounds.append([float(self.alpha.minimum_bound.text()),
+#                        float(self.alpha.maximum_bound.text())])
+
+#         specified_parameter_indices = [i for i, x in enumerate(self.DV_flags) if x==False]
+#         #Remove all bounds that are already specified.
+#         for index in sorted(specified_parameter_indices,reverse=True):
+#             del bounds[index]
+
+
+#         return bounds
+
+#     def getSpecifiedValues(self):
+#         #get the values that were specified to be constrained
+#         ## MAYBE: ONLY MAKE THE TEXT BOXES ABLE TO BE ACCESSED WHEN THE ASSOCIATED
+#         ## CHECKBOX IS CLICKED
+#         self.known_values = {}
+
+#         DV_order = ['E_M','E_A', 'M_s', 'M_s-M_f', 'A_s', 'A_f - A_s', 'C_M', 'C_A', 'H_min',
+#                     'H_max - H_min', 'k', 'n_1', 'n_2', 'n_3', 'n_4', 'sig_crit', 'alpha']
+
+
+#         lineEdits = [
+#             self.E_M.specified_value,
+#             self.E_A.specified_value,
+#             self.M_s.specified_value,
+#             self.M_s_M_f.specified_value,
+#             self.A_s.specified_value,
+#             self.A_f_A_s.specified_value,
+#             self.C_M.specified_value,
+#             self.C_A.specified_value,
+#             self.H_min.specified_value,
+#             self.H_max_H_min.specified_value,
+#             self.k.specified_value,
+#             self.n_1.specified_value,
+#             self.n_2.specified_value,
+#             self.n_3.specified_value,
+#             self.n_4.specified_value,
+#             self.sigma_0.specified_value,
+#             self.alpha.specified_value,
+#             ]
+
+#         for counter in range(len(DV_order)):
+#             if DV_order[counter] == 'E_A' and self.flags['modulus_flag'] == True:
+#                 pass
+#             elif DV_order[counter] == 'C_A' and self.flags['slope_flag'] == True:
+#                 pass
+#             elif DV_order[counter] in ['n_2','n_3','n_4'] and self.flags['smooth_hardening_flag'] == True:
+#                 pass
+#             elif self.DV_flags[counter] == False:
+#                 self.known_values[DV_order[counter]] = float(lineEdits[counter].text())
+
+#         self.delta = float(self.delta.lineEdit.text())
+#         self.sigma_cal = float(self.sigma_cal.lineEdit.text())
+#         self.MVF_tol = float(self.MVF_tol.lineEdit.text())
+
+#         self.num_gens = float(self.gen_line.text())
+#         self.pop_size = float(self.pop_line.text())
+#         self.num_iters = float(self.iter_line.text())
+
+#     def create_label(
+#             self,
+#             label_object,
+#             label_text,
+#             font_size,
+#             font_weight,
+#             parameter_label_width,
+#             parameter_label_height,
+#             ):
+
+
+#         size_policy = self.set_size_policy(
+#             label_object
+#             )
+#         label_object.setSizePolicy(
+#             size_policy
+#             )
+
+#         label_object.setMinimumSize(
+#             QtCore.QSize(
+#                 parameter_label_width,
+#                 parameter_label_height
+#                 )
+#             )
+#         label_object.setMaximumSize(
+#             QtCore.QSize(
+#                 parameter_label_width,
+#                 parameter_label_height
+#                 )
+#             )
+
+#         label_object.setWordWrap(True)
+#         label_object.setText(label_text)
+
+#         font = self.create_font(
+#             font_size,
+#             font_weight,
+#             bold_flag = True
+#             )
+
+#         label_object.setFont(font)
+#         label_object.setAlignment(QtCore.Qt.AlignCenter)
+#         label_object.setObjectName(label_text)
+
+
+
+#     def create_material_property(
+#             self,
+#             property_object,
+#             name,
+#             units,
+#             parameter_label_width,
+#             parameter_label_height,
+#             line_edit_width,
+#             line_edit_height,
+#             small_spacer,
+#             regular_spacer):
+
+#         property_object.setObjectName(str(name))
+
+#         property_object.label = textToLatex(
+#             r"$"+name+"$ [$\mathrm{"+str(units)+"}$]:",
+#             parameter_label_width,
+#             parameter_label_height,
+#             self.centralwidget
+#             )
+#         property_object.addWidget(property_object.label)
+
+#         property_object.addItem(small_spacer)
+
+#         property_object.minimum_bound = QtWidgets.QLineEdit(self.centralwidget)
+
+#         size_policy = self.set_size_policy(
+#             property_object.minimum_bound
+#             )
+#         property_object.minimum_bound.setSizePolicy(
+#             size_policy
+#             )
+#         property_object.minimum_bound.setMinimumSize(
+#             QtCore.QSize(line_edit_width, line_edit_height)
+#             )
+#         property_object.minimum_bound.setMaximumSize(
+#             QtCore.QSize(line_edit_width, line_edit_height)
+#             )
+#         property_object.setObjectName(str(name)+"_minimum_bound")
+#         property_object.addWidget(property_object.minimum_bound)
+
+#         property_object.addItem(regular_spacer)
+
+#         property_object.maximum_bound = QtWidgets.QLineEdit(self.centralwidget)
+
+#         size_policy = self.set_size_policy(
+#             property_object.maximum_bound
+#             )
+#         property_object.maximum_bound.setSizePolicy(
+#             size_policy
+#             )
+#         property_object.maximum_bound.setMinimumSize(
+#             QtCore.QSize(line_edit_width, line_edit_height)
+#             )
+#         property_object.maximum_bound.setMaximumSize(
+#             QtCore.QSize(line_edit_width, line_edit_height)
+#             )
+#         property_object.setObjectName(str(name)+"_maximum_bound")
+#         property_object.addWidget(property_object.maximum_bound)
+
+#         property_object.addItem(regular_spacer)
+
+#         property_object.checkBox = QtWidgets.QCheckBox(self.centralwidget)
+#         property_object.checkBox.stateChanged.connect(self.uncheck)
+#         property_object.checkBox.setText("")
+#         property_object.checkBox.setObjectName(str(name)+"_checkBox")
+#         property_object.addWidget(property_object.checkBox)
+
+#         property_object.addItem(regular_spacer)
+
+#         property_object.specified_value = QtWidgets.QLineEdit(
+#             self.centralwidget
+#             )
+#         property_object.specified_value.setSizePolicy(
+#             size_policy
+#             )
+#         property_object.specified_value.setMinimumSize(
+#             QtCore.QSize(line_edit_width, line_edit_height)
+#             )
+#         property_object.specified_value.setMaximumSize(
+#             QtCore.QSize(line_edit_width, line_edit_height)
+#             )
+#         property_object.setObjectName(str(name)+"_specified_value")
+#         property_object.addWidget(property_object.specified_value)
+
+#         property_object.addItem(regular_spacer)
+#         property_object.specified_value.setEnabled(False)
+#         property_object.checkBox.toggled.connect(
+#             property_object.specified_value.setEnabled
+#             )
+#         property_object.checkBox.toggled.connect(
+#             property_object.maximum_bound.setDisabled
+#             )
+#         property_object.checkBox.toggled.connect(
+#             property_object.minimum_bound.setDisabled
+#             )
+
+#     def create_constraint(
+#             self,
+#             constraint_object,
+#             name,
+#             text,
+#             text_width,
+#             text_height,
+#             font_weight,
+#             font_size,
+#             latex_flag = False):
+
+#         constraint_object.setContentsMargins(-1, -1, 5, -1)
+#         constraint_object.setObjectName(name)
+
+#         if latex_flag == False:
+#             pass
+#         else:
+#             constraint_object.label = textToLatex(
+#                 text,
+#                 text_width,
+#                 text_height,
+#                 self.centralwidget
+#                 )
+
+#         font = self.create_font(
+#             font_size,
+#             font_weight,
+#             bold_flag = True,
+#             )
+
+#         constraint_object.label.setFont(font)
+#         constraint_object.label.setAlignment(
+#             QtCore.Qt.AlignRight|QtCore.Qt.AlignTrailing|QtCore.Qt.AlignVCenter
+#             )
+#         constraint_object.label.setObjectName(name+"_label")
+#         constraint_object.addWidget(constraint_object.label)
+
+#         constraint_object.checkbox = QtWidgets.QCheckBox(self.centralwidget)
+#         constraint_object.checkbox.stateChanged.connect(self.uncheck)
+#         constraint_object.checkbox.setText("")
+#         constraint_object.checkbox.setObjectName(name+"_flag")
+#         constraint_object.addWidget(constraint_object.checkbox)
+
+#     def create_parameter(
+#             self,
+#             parameter_object,
+#             name,
+#             text,
+#             text_width,
+#             text_height,
+#             text_size,
+#             text_weight,
+#             lineEdit_width,
+#             lineEdit_height,
+#             latex_flag = False,
+#             ):
+
+#         parameter_object.setContentsMargins(-1, -1, 5, -1)
+#         parameter_object.setObjectName(name)
+
+
+#         # self.label_34 = QtWidgets.QLabel(self.centralwidget)
+#         if latex_flag == True:
+#             parameter_object.label = textToLatex(
+#                 text,
+#                 text_width,
+#                 text_height,
+#                 self.centralwidget
+#                 )
+#         else:
+#             parameter_object.label = QtWidgets.QLabel(self.centralwidget)
+#             parameter_object.label.setText(text)
+#             pass
+
+#         size_policy = self.set_size_policy(
+#             parameter_object.label
+#             )
+#         parameter_object.label.setSizePolicy(
+#             size_policy
+#             )
+
+#         parameter_object.label.setMinimumSize(
+#             QtCore.QSize(
+#                 text_width,
+#                 0
+#                 )
+#             )
+#         parameter_object.label.setMaximumSize(
+#             QtCore.QSize(
+#                 text_width,
+#                 16777215
+#                 )
+#             )
+
+#         font = self.create_font(
+#             text_size,
+#             text_weight,
+#             bold_flag = True
+#             )
+
+#         parameter_object.label.setFont(font)
+
+#         parameter_object.label.setAlignment(QtCore.Qt.AlignRight|QtCore.Qt.AlignTrailing|QtCore.Qt.AlignVCenter)
+
+#         parameter_object.label.setObjectName(name+"_label")
+#         parameter_object.addWidget(parameter_object.label)
+
+
+#         parameter_object.lineEdit = QtWidgets.QLineEdit(self.centralwidget)
+#         size_policy = self.set_size_policy(
+#             parameter_object.lineEdit
+#             )
+#         parameter_object.lineEdit.setSizePolicy(
+#             size_policy
+#             )
+
+#         parameter_object.lineEdit.setMinimumSize(QtCore.QSize(
+#             lineEdit_width,
+#             lineEdit_height
+#             )
+#             )
+#         parameter_object.lineEdit.setMaximumSize(
+#             QtCore.QSize(
+#                 lineEdit_width,
+#                 lineEdit_height
+#                 )
+#             )
+#         parameter_object.lineEdit.setObjectName(name+"_lineEdit")
+#         parameter_object.addWidget(parameter_object.lineEdit)
+
+
+
+
+#     def set_size_policy(
+#             self,
+#             lineEdit,
+#             ):
+#         sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
+#         sizePolicy.setHorizontalStretch(0)
+#         sizePolicy.setVerticalStretch(0)
+#         sizePolicy.setHeightForWidth(
+#             lineEdit.sizePolicy().hasHeightForWidth()
+#             )
+#         return sizePolicy
+
+#     def create_font(
+#             self,
+#             font_size,
+#             font_weight,
+#             bold_flag = False,
+#             ):
+#         font = QtGui.QFont()
+#         font.setPointSize(font_size)
+#         font.setBold(bold_flag)
+#         font.setWeight(font_weight)
+
+#         return font
+
+
+    # # Questionable if I need this.
+    # @pyqtSlot()
+    # def on_click(self):
+    #     print("\n")
+    #     for currentQTableWidgetItem in self.tableWidget.selectedItems():
+    #         print(currentQTableWidgetItem.row(), currentQTableWidgetItem.column(), currentQTableWidgetItem.text())
+
+
+
+#%% Calibration window widget
+class CalibrationProgressWidget(QtWidgets.QWidget):
+    def __init__(self):
+        super(QtWidgets.QWidget,self).__init__()
+
+        rc.update({'font.size': 18})
+
+        # SETUP
+        #self.setWindowTitle('Calibration Window')
+        self.main_layout = QGridLayout(self)
+        self.setLayout(self.main_layout)
+
+        # CANVASES/GRAPHS
+        self.opt_progress_graph = Figure()
+        self.opt_progress_canvas = FigureCanvas(self.opt_progress_graph)
+        self.opt_progress_toolbar = NavigationToolbar(self.opt_progress_canvas, self)
+
+        self.dv_vals_graph = Figure()
+        self.dv_vals_canvas = FigureCanvas(self.dv_vals_graph)
+        self.dv_vals_toolbar = NavigationToolbar(self.dv_vals_canvas, self)
+        keys = ['E_M', 'E_A', 'M_s', 'M_s - M_f', 'A_s', 'A_f - A_s', 'C_M', 'C_A', 'H_min', 'H_max - H_min', 'k',
+                'n_1', 'n_2', 'n_3', 'n_4', 'sig_crit', 'alpha']
+        y = np.zeros(len(keys))
+        self.ax_dv_vals = self.dv_vals_graph.add_subplot(111)
+        self.bars = self.ax_dv_vals.bar(keys, y, width=0.5)
+        self.ax_dv_vals.set_ylim([-0.1, 1.1])
+        self.ax_dv_vals.text(0.05,-0.066,'Lower bound')
+        self.ax_dv_vals.text(0.05,1.033,'Upper bound')
+        self.ax_dv_vals.set_ylabel('Normalized material property, -')
+        self.ax_dv_vals.axhline(0,color='black',linestyle='--')
+        self.ax_dv_vals.axhline(1,color='black',linestyle='--')
+
+        self.dv_vals_graph.autofmt_xdate(rotation=45, ha='right')
+
+        self.temp_strain_graph = Figure()
+        self.ax_temp_strain = self.temp_strain_graph.add_subplot(111)
+        self.temp_strain_canvas = FigureCanvas(self.temp_strain_graph)
+        self.temp_strain_toolbar = NavigationToolbar(self.temp_strain_canvas, self)
+        self.ax_temp_strain.set_autoscaley_on(True)
+        self.ax_temp_strain.set_xlabel('Temperature, K')
+        self.ax_temp_strain.set_ylabel('Strain, mm/mm')
+        self.ax_temp_strain.grid()
+
+        self.phase_diagram_graph = Figure()
+        self.ax_phase_digram = self.phase_diagram_graph.add_subplot(111)
+        self.phase_diagram_canvas = FigureCanvas(self.phase_diagram_graph)
+        self.phase_diagram_toolbar = NavigationToolbar(self.phase_diagram_canvas, self)
+
+        # ADDING TO WINDOW
+        self.main_layout.addWidget(self.opt_progress_canvas, 0, 0)
+        self.main_layout.addWidget(self.opt_progress_toolbar, 1, 0)
+        self.main_layout.addWidget(self.dv_vals_canvas, 2, 0)
+        self.main_layout.addWidget(self.dv_vals_toolbar, 3, 0)
+        self.main_layout.addWidget(self.temp_strain_canvas, 0, 1)
+        self.main_layout.addWidget(self.temp_strain_toolbar, 1, 1)
+        self.main_layout.addWidget(self.phase_diagram_canvas, 2, 1)
+        self.main_layout.addWidget(self.phase_diagram_toolbar, 3, 1)
+        self.show()
+
+
+    def run(self, bounds, knownValues, DV_flags):
+        QApplication.processEvents()
+        main(bounds, knownValues, DV_flags, self)
+
+
+    def updateOptProgress(self, gen, min_func_val, avg, std):
+        self.opt_progress_graph.clear()
+        ax = self.opt_progress_graph.add_subplot(111)
+        ax.plot(gen, min_func_val, color='red', label='Current best solution')
+        ax.plot(gen, avg, label='Average solution',color='black')
+        #ax.fill_between(gen, [x + y for x,y in zip(avg, std)], [x - y for x,y in zip(avg, std)],
+        #                                  color='#888888', alpha=0.2)
+        ax.legend(loc='best')
+        ax.set_xlim([0, max(gen) + 1])
+        ax.set_xlabel('Generation')
+        ax.set_ylabel('Error')
+        self.opt_progress_canvas.draw()
+        self.opt_progress_canvas.flush_events()
+
+
+    def updateDVVals(self, dv,DV_flags):
+        count = 0
+        for i, bar in enumerate(self.bars):
+            if DV_flags[i] == True:
+                bar.set_height(dv[count])
+                count +=1
+            else:
+                pass
+                # bar.set_height(1)
+                # bar.set_color('red')
+
+        self.dv_vals_canvas.draw()
+        self.dv_vals_canvas.flush_events()
+
+
+    def plotExperimental(self, x, y, numExps):
+        self.ax_temp_strain.clear()
+        self.ax_temp_strain.plot(x, y, 'bo',label='Experiment')
+        self.temp_strain_canvas.draw()
+        self.temp_strain_canvas.flush_events()
+
+        self.lines = [self.ax_temp_strain.plot([], [], 'r-')[0] for _ in range(numExps)]
+
+
+    def updateTempStrain(self, temp, strain, numExps):
+        for i in range(numExps):
+            self.lines[i].set_xdata(temp[i])
+            self.lines[i].set_ydata(strain[i])
+
+        self.lines[i].set(label='Model prediction')
+        self.ax_temp_strain.legend(loc='best')
+        self.ax_temp_strain.set_xlabel('Temperature, K',fontsize=18)
+        self.ax_temp_strain.set_ylabel('Strain, mm/mm',fontsize=18)
+
+        self.ax_temp_strain.relim()
+        self.ax_temp_strain.autoscale_view()
+
+        self.temp_strain_canvas.draw()
+        self.temp_strain_canvas.flush_events()
+
+
+    def updatePhaseDiagram(self, P, sigma_inp):
+        self.ax_phase_digram.clear()
+        plot_phase_diagram(P, sigma_inp, self.ax_phase_digram)
+
+        self.phase_diagram_canvas.draw()
+        self.phase_diagram_canvas.flush_events()
+
+
+
+
+
+
+# if __name__ == "__main__":
+#     import sys
+#     app = QtWidgets.QApplication(sys.argv)
+#     MainWindow = QtWidgets.QMainWindow()
+#     ex = App()
+#     #ui = Ui_MainWindow()
+#     #ui.setupUi(MainWindow)
+#     #MainWindow.show()
+#     sys.exit(app.exec_())
